requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.1
colorama==0.4.6
dataclasses-json==0.6.1
Django==4.2.6
frozenlist==1.4.0
google-ai-generativelanguage==0.3.3
google-api-core==2.12.0
google-auth==2.23.3
google-generativeai==0.2.2
google-search-results==2.4.2
googleapis-common-protos==1.61.0
greenlet==3.0.1
grpcio==1.59.0
grpcio-status==1.59.0
idna==3.4
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.325
langchainhub==0.1.13
langsmith==0.0.53
markdown-it-py==3.0.0
marshmallow==3.20.1
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.1
packaging==23.2
pandas==2.1.2
proto-plus==1.22.3
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.16.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
rich==13.6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.22
sqlparse==0.4.4
tenacity==8.2.3
tqdm==4.66.1
types-requests==2.31.0.10
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
yarl==1.9.2
